backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.1
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
gunicorn==22.0.0
boto3==1.34.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
transformers==4.40.0
torch==2.3.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
scipy==1.13.0
hypothesis==6.103.1
pytest==8.2.2
pytest-flask==1.3.0
